[NO-ISSUE] refactor: 상수 추출 및 컨벤션 적용
