--- src/main/java/com/sds/actlongs/domain/video/respository/VideoRepository.java
@@ -1,9 +1,0 @@
-package com.sds.actlongs.domain.video.respository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.sds.actlongs.domain.video.entity.Video;
-
-public interface VideoRepository extends JpaRepository<Video, Long> {
-
-}
