+++ src/main/java/com/sds/actlongs/util/uuid/UuidGenerateImpl.java
@@ -1,5 +1,7 @@
 package com.sds.actlongs.util.uuid;
 
+import static com.sds.actlongs.util.Constants.*;
+
 import java.util.UUID;
 
 import org.springframework.stereotype.Component;
@@ -9,7 +11,7 @@ public class UuidGenerateImpl implements UuidGenerate {
 
 	@Override
 	public String getVodUuid() {
-		return UUID.randomUUID().toString().replaceAll("-", "");
+		return UUID.randomUUID().toString().replaceAll(HYPHEN, EMPTY);
 	}
 
 }
